Add files via upload
The main aim of this project is to predict whether a person will leave the Bank services, i.e he/she will churn or not using Machine Learning and Auto Machine Learning Techniques. In general, Churn is expressed as a degree of customer inactivity or disengagement, observed over a given time. This manifests within the data in various forms such as the recency of account actions or change in the account balance.
This project aims to accomplish the following:
-Visualizing which factors contribute to customer churn.
-Build a prediction model that will classify whether a customer is going to churn or not.
-Based on the model performance, choosing a model that will attach a probability to the churn to make it easier for the customer service to target low hanging fruits in their efforts to prevent churn.